pyproject.toml
[project]
name = "j7-c-usb-meter"
version = "0.1.0"
description = "J7-C USB Tester Data Logger"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bleak>=2.1.1",
    "pyserial>=3.5",
]

[project.scripts]
j7-c-usb-tester = "j7_c_usb_tester.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/j7_c_usb_tester"]
